sharing prototype mostly implemented
