+++ app/src/androidTest/java/com/github/scribeWizTeam/scribewiz/AuthActivityTest.kt
@@ -9,6 +9,7 @@ import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.intent.Intents.intended
 import androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent
 import com.firebase.ui.auth.KickoffActivity
+import com.github.scribeWizTeam.scribewiz.Activities.FirebaseUIActivity
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
