+++ app/src/androidTest/java/com/github/scribeWizTeam/scribewiz/MainActivityTest.kt
@@ -2,16 +2,13 @@ package com.github.scribeWizTeam.scribewiz
 
 import org.junit.Rule
 
-import androidx.test.core.app.ActivityScenario
-import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.*
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Before
+import com.github.scribeWizTeam.scribewiz.Activities.MainActivity
 import org.junit.Test
 import org.junit.runner.RunWith
