+++ app/src/main/java/com/github/scribeWizTeam/scribewiz/Activities/NavigationActivity.kt
@@ -1,4 +1,4 @@
-package com.github.scribeWizTeam.scribewiz
+package com.github.scribeWizTeam.scribewiz.Activities
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -8,6 +8,8 @@ import androidx.appcompat.widget.Toolbar
 import androidx.core.view.GravityCompat
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.fragment.app.Fragment
+import com.github.scribeWizTeam.scribewiz.Fragments.*
+import com.github.scribeWizTeam.scribewiz.R
 import com.google.android.material.navigation.NavigationView
 
 
@@ -71,7 +73,7 @@ class NavigationActivity : AppCompatActivity(), NavigationView.OnNavigationItemS
             R.id.nav_library -> showFragment(LibraryFragment(0))
             R.id.nav_rec -> showFragment(RecFragment(0))
             R.id.nav_settings -> showFragment(SettingsFragment(0))
-            R.id.nav_share -> ShareFragment(0).shareMidiFile("/storage/emulated/0/Download/1.mid", this)
+            R.id.nav_share -> ShareFragment(0).shareMidiFile("/storage/emulated/0/Android/data/com.github.scribeWizTeam.scribewiz/cache/recording.3gp", this)
         }
         //close the navigation drawer
         drawer.closeDrawer(GravityCompat.START)
