+++ app/src/main/java/com/github/scribeWizTeam/scribewiz/Fragments/HelpFragment.kt
@@ -1,10 +1,11 @@
-package com.github.scribeWizTeam.scribewiz
+package com.github.scribeWizTeam.scribewiz.Fragments
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.github.scribeWizTeam.scribewiz.R
 
 class HelpFragment(contentLayoutId: Int) : Fragment(contentLayoutId) {
