+++ app/src/main/res/layout/activity_navigation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity"
+    tools:context=".Activities.MainActivity"
     android:id="@+id/drawer_layout"
     android:fitsSystemWindows="true"
     tools:layout_editor_absoluteX="0dp"
