+++ app/src/main/res/xml/provider_paths.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="utf-8"?>
+<paths>
+    <paths>
+        <external-path
+            name="external"
+            path="." />
+        <external-files-path
+            name="external_files"
+            path="." />
+        <cache-path
+            name="cache"
+            path="." />
+        <external-cache-path
+            name="external_cache"
+            path="." />
+        <files-path
+            name="files"
+            path="." />
+    </paths>
+</paths>
